### 2025-08-21 #8848 @editorialbot
[PRE REVIEW]: Mastodon.py: A Python library for the Mastodon API
**Submitting author:** @halcy (<a href="http://orcid.org/0000-0001-9370-9243">Lorenz Diener</a>)
**Repository:** https://github.com/halcy/Mastodon.py
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** v2.1.2
