### 2025-08-21 #8848 @danielskatz
👋 @halcy - thanks for your submission.

Can you please add countries to your affiliations in your .md file?

Please also add a References header to the end of your paper so that this appears before the generated reference in the pdf.

See the [example paper](https://joss.readthedocs.io/en/latest/example_paper.html). In addition, you could work on the possibly missing DOI that editorialbot suggests, but note that it may be incorrect. Please feel free to make changes to your .bib file, then use the command `@editorialbot check references` to check again, and the command `@editorialbot generate pdf` after making changes to the .md file or when the references are right to make a new PDF. editorialbot commands need to be the first entry in a new comment.

Otherwise, I think this is ready to get started, but we don't currently have an appropriate editor available, so I'm going to put this on our waiting list for now.
