### 2025-08-21 5df4839
Update paper.bib
Add forgotten DOI
