+++ paper/paper.bib
@@ -54,6 +54,7 @@
   pages={357--380},
   year={2024},
   publisher={Springer},
+  doi={10.1007/s11616-024-00855-6},
 }
 
 @misc{scidsg2024mastodonScheduler,
