### 2025-08-21 b95a846
Update paper.md
Add # References
