### 2025-08-21 a906b30
Update paper.md
Add affiliation locations
