### 2025-08-21 #8848 @halcy
@danielskatz Thank you for the extremely quick first response, and what a convenient bot, I want that to check all my papers for missing DOIs. I've adjusted the .md and .bib .

Reviewer suggestions: Unfortunately, I can't really think of anyone offhand. But if I do, I will let you know. What kind of background would make for a good reviewer for JOSS / this type of paper?

Also, just for clarification and matching up with the commit list: I am Lorenz "halcy" Diener (and, as you can see, sometimes sloppy with setting my git identity), my coauthor is Corentin "codl" Delcourt.
