### 2025-08-22 1b2ac8b
Update and rename DEVELOPMENT.md to CONTRIBUTING.md
