### 2025-08-22 50eca6b
Create SECURITY.md
