### 2025-08-22 6ee8169
Update README.rst
