### 2025-08-26 48ad964
fix some doc issues, add run_mastodon_for_tests.sh
