### 2025-09-01 c7d2b7e
fix #420 - warnings on some internal __instance calls
