### 2025-09-01 53e9a33
fix "everything" docs page
