+++ docs/conf.py
@@ -68,7 +68,7 @@ author = u'Lorenz Diener'
 # The short X.Y version.
 version = u'2.1'
 # The full version, including alpha/beta/rc tags.
-release = u'2.1.2'
+release = u'2.1.3'
 
 # The language for content autogenerated by Sphinx. Refer to documentation
 # for a list of supported languages.
