+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend= "setuptools.build_meta"
 
 [project]
 name = "Mastodon.py"
-version = "2.1.2"
+version = "2.1.3"
 description = "Python wrapper for the Mastodon API"
 readme = "README.rst"
 authors = [
