### 2025-09-06 #8946 @bkrayfield
👋 Hello Everyone,

The editorial bot has outlined the instructions clearly above. We can now begin the review process—please be sure to generate your checklist and link any comments back to this thread.

I’ll give you space to work here, but just a reminder: we ask that reviews be completed within six weeks, while recognizing that some back-and-forth discussion may naturally occur along the way.
