### 2025-09-10 d00c131
Specify `client_name` type as `str`
