### 2025-09-11 0498b70
Update CHANGELOG for versioning and v2.1.4 changes
