### 2025-09-22 2d0d11b
Use bounded type variables for pagination utility methods
