+++ mastodon/utility.py
@@ -147,7 +147,7 @@ class Mastodon(Internals):
     ###
     # Pagination
     ###
-    def fetch_next(self, previous_page: Union[PaginatableList[Entity], Entity, PaginationInfo]) -> Optional[Union[PaginatableList[Entity], Entity]]:
+    def fetch_next[T: Entity](self, previous_page: Union[PaginatableList[T], T, PaginationInfo]) -> Optional[Union[PaginatableList[T], T]]:
         """
         Fetches the next page of results of a paginated request. Pass in the
         previous page in its entirety, or the pagination information dict
@@ -195,7 +195,7 @@ class Mastodon(Internals):
         else:
             return self.__api_request(method, endpoint, params, override_type=response_type)
 
-    def fetch_previous(self, next_page: Union[PaginatableList[Entity], Entity, PaginationInfo]) -> Optional[Union[PaginatableList[Entity], Entity]]:
+    def fetch_previous[T: Entity](self, next_page: Union[PaginatableList[T], T, PaginationInfo]) -> Optional[Union[PaginatableList[T], T]]:
         """
         Fetches the previous page of results of a paginated request. Pass in the
         previous page in its entirety, or the pagination information dict
@@ -243,7 +243,7 @@ class Mastodon(Internals):
         else:
             return self.__api_request(method, endpoint, params, override_type=response_type)
 
-    def fetch_remaining(self, first_page: PaginatableList[Entity]) -> PaginatableList[Entity]:
+    def fetch_remaining[T: Entity](self, first_page: PaginatableList[T]) -> PaginatableList[T]:
         """
         Fetches all the remaining pages of a paginated request starting from a
         first page and returns the entire set of results (including the first page
@@ -281,7 +281,7 @@ class Mastodon(Internals):
         else:
             return None
 
-    def pagination_iterator(self, start_page: Union[PaginatableList[Entity], PaginationInfo], direction: str = "next", return_pagination_info: bool = False) -> Iterator[Entity]:
+    def pagination_iterator[T: Entity](self, start_page: Union[PaginatableList[T], PaginationInfo], direction: str = "next", return_pagination_info: bool = False) -> Iterator[T]:
         """
         Returns an iterator that will yield all entries in a paginated request,
         starting from the given start_page (can also be just the PaginationInfo, in which case the
