### 2025-09-22 29e9232
Use older `TypeVar()` syntax
