### 2025-09-28 e44c054
remove debug prints. fixes #425
