+++ tests/cassettes_entity_tests/test_entity_mediaattachmentmetadatacontainer.yaml
@@ -21,29 +21,29 @@ interactions:
         false, "spoiler_text": "", "visibility": "private", "language": "de", "uri":
         "https://icosahedron.website/users/halcy/statuses/110447012380248325", "url":
         "https://icosahedron.website/@halcy/110447012380248325", "replies_count":
-        0, "reblogs_count": 0, "favourites_count": 0, "edited_at": null, "favourited":
-        false, "reblogged": false, "muted": false, "bookmarked": false, "content":
-        "<p><span class=\"h-card\"><a href=\"https://mastodon.social/@halcy\" class=\"u-url
-        mention\" rel=\"nofollow noopener\" target=\"_blank\">@<span>halcy@mastodon.social</span></a></span>
+        0, "reblogs_count": 0, "favourites_count": 0, "quotes_count": 0, "edited_at":
+        null, "favourited": false, "reblogged": false, "muted": false, "bookmarked":
+        false, "content": "<p><span class=\"h-card\"><a href=\"https://mastodon.social/@halcy\"
+        class=\"u-url mention\" rel=\"nofollow noopener\" target=\"_blank\">@<span>halcy@mastodon.social</span></a></span>
         audio post, but remote</p>", "filtered": [], "reblog": null, "account": {"id":
         "23972", "username": "halcy", "acct": "halcy@icosahedron.website", "display_name":
         "halcy\u200b :icosahedron:", "locked": false, "bot": false, "discoverable":
         true, "indexable": true, "group": false, "created_at": "2017-02-03T00:00:00.000Z",
-        "note": "<p>build. your. squid.</p><p>ideal for comfortable gameplay in relaxed
-        locations.</p><p>i work at, but do not speak for, microsoft.</p>", "url":
-        "https://icosahedron.website/@halcy", "uri": "https://icosahedron.website/users/halcy",
+        "note": "<p>i am exactly like all the other guys.</p><p>ideal for comfortable
+        gameplay in relaxed locations.</p><p>i work at, but do not speak for, microsoft.</p>",
+        "url": "https://icosahedron.website/@halcy", "uri": "https://icosahedron.website/users/halcy",
         "avatar": "https://files.mastodon.social/cache/accounts/avatars/000/023/972/original/9c900052b7f670db.jpg",
         "avatar_static": "https://files.mastodon.social/cache/accounts/avatars/000/023/972/original/9c900052b7f670db.jpg",
         "header": "https://files.mastodon.social/cache/accounts/headers/000/023/972/original/e5c0d647464517c2.png",
         "header_static": "https://files.mastodon.social/cache/accounts/headers/000/023/972/original/e5c0d647464517c2.png",
-        "followers_count": 3421, "following_count": 519, "statuses_count": 68338,
-        "last_status_at": "2025-08-15", "hide_collections": true, "emojis": [{"shortcode":
+        "followers_count": 3483, "following_count": 525, "statuses_count": 70985,
+        "last_status_at": "2025-11-06", "hide_collections": true, "emojis": [{"shortcode":
         "icosahedron", "url": "https://files.mastodon.social/cache/custom_emojis/images/000/002/116/original/e0d448640e549b87.png",
         "static_url": "https://files.mastodon.social/cache/custom_emojis/images/000/002/116/static/e0d448640e549b87.png",
         "visible_in_picker": true}], "fields": [{"name": "Web", "value": "<a href=\"https://halcy.de/\"
[…]
