+++ tests/cassettes/test_admin_domain_blocks.yaml
@@ -20,27 +20,27 @@ interactions:
     uri: http://localhost:3000/api/v1/admin/domain_blocks/
   response:
     body:
-      string: '{"id":"1","domain":"chitter.xyz","digest":"06f1118b8a3b4cb0b92683b284f4c0c4a03e43b151247acda54c054fcb0d5174","created_at":"2025-08-17T18:41:57.663Z","severity":"suspend","reject_media":false,"reject_reports":false,"private_comment":null,"public_comment":"sicko
+      string: '{"id":"1","domain":"chitter.xyz","digest":"06f1118b8a3b4cb0b92683b284f4c0c4a03e43b151247acda54c054fcb0d5174","created_at":"2025-11-06T21:19:01.325Z","severity":"suspend","reject_media":false,"reject_reports":false,"private_comment":null,"public_comment":"sicko
         behaviour","obfuscate":false}'
     headers:
-      Content-Length:
-      - '290'
       cache-control:
       - private, no-store
+      content-length:
+      - '290'
       content-security-policy:
       - default-src 'none'; frame-ancestors 'none'; form-action 'none'
       content-type:
       - application/json; charset=utf-8
       etag:
-      - W/"4101a340928e35148c4d73e3fdc2bacb"
+      - W/"be881386f14e67515f98d9796c956599"
       referrer-policy:
       - strict-origin-when-cross-origin
       server-timing:
-      - cache_read.active_support;dur=0.03, sql.active_record;dur=20.67, cache_generate.active_support;dur=0.92,
+      - cache_read.active_support;dur=0.03, sql.active_record;dur=23.30, cache_generate.active_support;dur=1.16,
         cache_write.active_support;dur=0.08, instantiation.active_record;dur=0.23,
-        start_processing.action_controller;dur=0.00, start_transaction.active_record;dur=0.01,
-        transaction.active_record;dur=11.94, render.active_model_serializers;dur=0.25,
-        process_action.action_controller;dur=107.78
+        start_processing.action_controller;dur=0.01, start_transaction.active_record;dur=0.01,
+        transaction.active_record;dur=10.92, render.active_model_serializers;dur=0.34,
+        process_action.action_controller;dur=90.71
       vary:
       - Authorization, Origin
       x-content-type-options:
@@ -54,11 +54,11 @@ interactions:
       x-ratelimit-remaining:
       - '299'
       x-ratelimit-reset:
-      - '2025-08-17T18:45:00.650210Z'
+      - '2025-11-06T21:20:00.308759Z'
       x-request-id:
-      - 45625112-0c7a-48c1-a72d-6e82eda551ed
+      - 9d01d0ef-c310-4fbb-8dd3-9b6670ff801d
       x-runtime:
-      - '0.130355'
+      - '0.118455'
       x-xss-protection:
       - '0'
     status:
@@ -81,28 +81,28 @@ interactions:
     uri: http://localhost:3000/api/v1/admin/domain_blocks/
   response:
     body:
-      string: '[{"id":"1","domain":"chitter.xyz","digest":"06f1118b8a3b4cb0b92683b284f4c0c4a03e43b151247acda54c054fcb0d5174","created_at":"2025-08-17T18:41:57.663Z","severity":"suspend","reject_media":false,"reject_reports":false,"private_comment":null,"public_comment":"sicko
+      string: '[{"id":"1","domain":"chitter.xyz","digest":"06f1118b8a3b4cb0b92683b284f4c0c4a03e43b151247acda54c054fcb0d5174","created_at":"2025-11-06T21:19:01.325Z","severity":"suspend","reject_media":false,"reject_reports":false,"private_comment":null,"public_comment":"sicko
         behaviour","obfuscate":false}]'
     headers:
-      Content-Length:
-      - '292'
[…]
