+++ tests/cassettes/test_favourites.yaml
@@ -18,10 +18,10 @@ interactions:
     body:
       string: '[]'
     headers:
-      Content-Length:
-      - '2'
       cache-control:
       - private, no-store
+      content-length:
+      - '2'
       content-security-policy:
       - default-src 'none'; frame-ancestors 'none'; form-action 'none'
       content-type:
@@ -31,10 +31,10 @@ interactions:
       referrer-policy:
       - strict-origin-when-cross-origin
       server-timing:
-      - cache_read.active_support;dur=0.05, sql.active_record;dur=1.42, cache_generate.active_support;dur=2.68,
-        cache_write.active_support;dur=0.12, instantiation.active_record;dur=0.15,
+      - cache_read.active_support;dur=0.05, sql.active_record;dur=1.95, cache_generate.active_support;dur=3.86,
+        cache_write.active_support;dur=0.13, instantiation.active_record;dur=0.25,
         start_processing.action_controller;dur=0.00, render.active_model_serializers;dur=0.04,
-        process_action.action_controller;dur=30.23
+        process_action.action_controller;dur=32.72
       vary:
       - Authorization, Origin
       x-content-type-options:
@@ -48,11 +48,11 @@ interactions:
       x-ratelimit-remaining:
       - '299'
       x-ratelimit-reset:
-      - '2025-08-17T18:50:00.600946Z'
+      - '2025-11-06T21:25:00.342564Z'
       x-request-id:
-      - 52558e53-595a-47ff-a36b-5557ba14501b
+      - a63d640d-5b3f-4cbf-aa36-3a03d6d85d19
       x-runtime:
-      - '0.051202'
+      - '0.054754'
       x-xss-protection:
       - '0'
     status:

[…]
