### 2025-11-06 939c750
Update pyproject.toml
Switch grapheme dependency to graphemeu, which is maintained
