+++ pyproject.toml
@@ -38,7 +38,7 @@ blurhash = [
     'blurhash>=1.1.4',
 ]
 grapheme = [
-    'grapheme>=0.6.0',
+    'graphemeu>=0.8.1',
 ]
 test = [
     'pytest',
