### 2025-11-06 505a45e
Update pyproject.toml
Change minimum to 0.7.2
