+++ pyproject.toml
@@ -38,7 +38,7 @@ blurhash = [
     'blurhash>=1.1.4',
 ]
 grapheme = [
-    'graphemeu>=0.8.1',
+    'graphemeu>=0.7.2',
 ]
 test = [
     'pytest',
