### 2025-11-08 1124a5a
Skip status length test on 3.7
