+++ tests/test_status_length.py
@@ -1,5 +1,6 @@
 
 import pytest
+import sys
 
 from  mastodon import Mastodon
 
@@ -54,5 +55,9 @@ TEST_CASES = [
 
 @pytest.mark.parametrize("text,expected", TEST_CASES)
 def test_get_status_length_against_ground_truth(text, expected):
+    # skip if python version is less than or equal 3.7
+    if sys.version_info <= (3, 7):
+        pytest.skip("Python version is less than or equal to 3.7")
+
     assert Mastodon.get_status_length(text) == expected
     assert Mastodon.get_status_length(text, "what") == expected + 4
