+++ tests/test_status_length.py
@@ -58,6 +58,6 @@ def test_get_status_length_against_ground_truth(text, expected):
     # skip if python version is less than or equal 3.7
     if sys.version_info <= (3, 7):
         pytest.skip("Python version is less than or equal to 3.7")
-
-    assert Mastodon.get_status_length(text) == expected
-    assert Mastodon.get_status_length(text, "what") == expected + 4
+    else:
+        assert Mastodon.get_status_length(text) == expected
+        assert Mastodon.get_status_length(text, "what") == expected + 4
