### 2025-12-05 2ac51ca
Fix UnboundLocalError in streaming API URL parsing
