### 2026-03-01 d50b555
Fix "Invalid event" error when server sent heartbeats
