### 2026-03-30 #8946 @danielskatz
👋 @halcy - I've proofread the paper, and suggest the changes in https://github.com/halcy/Mastodon.py/pull/440 - please merge this, or let me know what you disagree with.
