### 2026-03-30 #8946 @halcy
@danielskatz One question before I look into archival services - should this archive include the joss paper, e.g. should I merge it to master, or can / should it continue to live in the branch?
