### 2026-03-30 #8946 @danielskatz
Either is fine. Since the paper will be archived as PDF and XML, the source is not required to be in the archived repo.
