### 2026-04-12 #8946 @halcy
@danielskatz Alright, I think I have completed evertything:

* Tagged version: **v2.2.0**: https://github.com/halcy/Mastodon.py/releases/tag/v2.2.0
* This version, archived on figshare, DOI **10.6084/m9.figshare.31989372**: https://figshare.com/articles/software/Mastodon_py_A_Python_library_for_the_Mastodon_API/31989372?file=63653637
