### 2026-04-12 #8946 @danielskatz
Congratulations to @halcy (Lorenz Diener) and co-author on your JOSS publication!!

And thanks to @professornaite and @dataspider for reviewing!
JOSS completely depends on volunteers - we couldn't do this without you
