### 2026-04-12 #8946 @halcy
Wow, that was fast! I've pushed out a new minor version of our software package to update the docs to point to the new JOSS publication, too already.

Thank you again to everyone for the reviewing and editing!
